requirements.txt
flask==2.3.2
twilio==8.5.0
SpeechRecognition==3.10.0
pyttsx3==2.90
pygame==2.5.0
requests==2.31.0
python-dotenv==1.0.0